allow result cache key to take field parameter
This is needed in order for the Field Extensions to produce a valid unique cache key.

When using Field Extensions the context doens't store the path the same way, however the it has information about the field itself, so this can be passed on to the Key class in order to compute a cache key with correct values.